requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
prophet
requests
apscheduler
python-dotenv
pydantic